requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
Click==7.0
coverage==4.5.2
cryptography==2.5
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
nose==1.3.7
psycopg2-binary==2.7.7
pycparser==2.19
PyJWT==1.7.1
six==1.12.0
SQLAlchemy==1.2.17
Werkzeug==0.14.1
